chore: move rust bindings file
