--- Cargo.toml
@@ -1,17 +1,0 @@
-[package]
-name = "evacuation"
-version = "0.1.0"
-edition = "2021"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-cli = { path = "src/cli" }
-configuration = { path = "src/configuration" }
-json_object = { path = "src/json_object" }
-bim_cli = { path = "src/bim_cli" }
-bim_configure = { path = "src/bim_configure" }
-bim_json_object = { path = "src/bim_json_object" }
-
-[build-dependencies]
-cc = { version = "1.0", features = ["parallel"] }
